Improved browser wallet connect/disconnect flow.
