+++ packages/core/src/lib/constants.ts
@@ -1,4 +1,5 @@
 export const PACKAGE_NAME = "near-wallet-selector";
-export const LOCAL_STORAGE_SELECTED_WALLET_ID = `selectedWalletId`;
+export const SELECTED_WALLET_ID = `selectedWalletId`;
+export const PENDING_SELECTED_WALLET_ID = `selectedWalletId:pending`;
 
 export const DEFAULT_DERIVATION_PATH = "44'/397'/0'/0'/1'";
